+++ src/util/const.ts
@@ -1,4 +1,5 @@
 import { Network } from '../types/types';
+import { Network as NetworkFromEthers } from '@ethersproject/networks';
 
 export const DEFAULT_CONTRACT_ADDRESSES = 'DEFAULT_TOKENS';
 export const DEFAULT_ALCHEMY_API_KEY = 'demo';
@@ -45,7 +46,24 @@ export const EthersNetwork = {
   [Network.ARB_RINKEBY]: 'arbitrum-rinkeby',
   [Network.ARB_GOERLI]: 'arbitrum-goerli',
   [Network.MATIC_MAINNET]: 'matic',
-  [Network.MATIC_MUMBAI]: 'maticmum'
+  [Network.MATIC_MUMBAI]: 'maticmum',
+  [Network.ASTAR_MAINNET]: 'astar-mainnet'
+};
+
+/**
+ * Mapping of network names to their corresponding Ethers Network objects. These
+ * networks are not yet supported by Ethers and are listed here to be overriden
+ * in the provider.
+ */
+export const CustomNetworks: { [key: string]: NetworkFromEthers } = {
+  'arbitrum-goerli': {
+    chainId: 421613,
+    name: 'arbitrum-goerli'
+  },
+  'astar-mainnet': {
+    chainId: 592,
+    name: 'astar-mainnet'
+  }
 };
 
 export function noop(): void {